refactor, extract loadVersion() without changing behavior
